+++ lib/i18n-js/embed_fallback_translations_plugin.rb
@@ -6,21 +6,10 @@ module I18nJS
   class EmbedFallbackTranslationsPlugin < I18nJS::Plugin
     CONFIG_KEY = :embed_fallback_translations
 
-    # This method must set up the basic plugin configuration, like adding the
-    # config's root key in case your plugin accepts configuration (defined via
-    # the config file).
-    #
-    # If you don't add this key, the linter will prevent non-default keys from
-    # being added to the configuration file.
     def self.setup
       I18nJS::Schema.root_keys << CONFIG_KEY
     end
 
-    # In case your plugin accepts configuration, this is where you must validate
-    # the configuration, making sure only valid keys and type is provided.
-    # If the configuration contains invalid data, then you must raise an
-    # exception using something like
-    # `raise I18nJS::Schema::InvalidError, error_message`.
     def self.validate_schema(config:)
       return unless config.key?(CONFIG_KEY)
 
@@ -33,14 +22,6 @@ module I18nJS
       schema.expect_enabled_config(CONFIG_KEY, plugin_config[:enabled])
     end
 
-    # This method is responsible for transforming the translations. The
-    # translations you'll receive may be already be filtered by other plugins
-    # and by the default filtering itself. If you need to access the original
-    # translations, use `I18nJS.translations`.
-    #
-    # Make sure you always check whether your plugin is active before
-    # transforming translations; otherwise, opting out transformation won't be
-    # possible.
     def self.transform(translations:, config:)
       return translations unless config.dig(CONFIG_KEY, :enabled)