+++ CHANGELOG.md
@@ -11,6 +11,12 @@ Prefix your message with one of the following:
 - [Security] in case of vulnerabilities.
 -->
 
+## Unreleased
+
+- [Added] Add `I18nJS::Plugin.after_export(files:)` method, that's called
+  whenever whenever I18nJS finishes exporting files. You can use it to further
+  process files, or generate new files based on the exported files.
+
 ## v4.1.0 - Dec 09, 2022
 
 - [Added] Parse configuration files as erb.